reorganization catalogs, some js code
