--- web/front/index_registration.html
@@ -1,39 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head> 
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Music Log In</title>
-  <link rel="stylesheet" href="style_login.css">
-</head>
-<body>
-  <div class="wrapper">
-    <form action="#">
-      <h2>Регистрация</h2>
-        <div class="input-field">
-        <input type="text" required>
-        <label>Введите адрес своей электронной почты</label>
-      </div>
-      <div class="input-field">
-        <input type="text" required>
-        <label>Введите логин</label>
-      </div>
-      <div class="input-field">
-        <input type="password" required>
-        <label>Введите пароль</label>
-      </div>
-      <div class="input-field">
-        <input type="password" required>
-        <label>Повторите пароль</label>
-      </div>
-      <div class="forget">
-        <label for="remember">
-          <input type="checkbox" id="remember">
-          <p>Запомнить меня</p>
-        </label>
-      </div>
-      <button type="submit">Зарегестрироваться</button>
-    </form>
-  </div>
-</body>
-</html>
