+++ web/front/src/app.jsx
@@ -1,0 +1,1 @@
+import { BrowserRouter, Routes, Route} from "react-router-dom"
