+++ web/front/src/main.jsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './css/index.css'
-import Entrance from './templates/entrance.jsx'
+import Entrance from './entrance.jsx'
 
 createRoot(document.getElementById('entrance')).render(
   <StrictMode>
