--- web/front/src/templates/index_login.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head> 
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Music Log In</title>
-  <link rel="stylesheet" href="style_login.css">
-</head>
-<body>
-  <div class="wrapper">
-    <form action="#">
-      <h2>Вход</h2>
-        <div class="input-field">
-        <input type="text" required>
-        <label>Введите адрес своей электронной почты</label>
-      </div>
-      <div class="input-field">
-        <input type="password" required>
-        <label>Введите пароль</label>
-      </div>
-      <div class="forget">
-        <label for="remember">
-          <input type="checkbox" id="remember">
-          <p>Запомнить меня</p>
-        </label>
-        <a href="#">Забыли пароль?</a>
-      </div>
-      <button type="submit">Войти</button>
-      <div class="register">
-        <p>Вы не зарегестрированны? <a href="index_registration.html">Регистрация</a></p>
-      </div>
-    </form>
-  </div>
-</body>
-</html>
