+++ web/front/src/templates/index_personal_account.html
@@ -5,14 +5,14 @@
     <title> Music home </title>
     <link rel="stylesheet" href="style_personal_account.css">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.2/css/all.min.css"/>
-     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-   </head>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  </head>
 <body>
   <header>
     <nav class="navbar">
       <div class="logo"><a href="#">MUSIC></a></div>
       <ul class="menu">
-        <li><a href="index_home.html">Главная</a></li>
+        <li><a href="/home">Главная</a></li>
         <li><a href="#">Личный кабинет</a></li>
       </ul>
       <div class="buttons">
