+++ web/front/src/templates/register.jsx
@@ -1,22 +1,28 @@
 import '../css/register.css'
+import {Helmet} from 'react-helmet'
+import RegisterUpl from '../js/register.js'
+
 
 function Register() {
   return (
     <>
     <div class="wrapper">
-      <form action="#">
+      <Helmet>
+        <title>Sign up | Ручей</title>
+      </Helmet>
+      <form action={RegisterUpl}>
         <h2>Sign up</h2>
         <div class="input-field">
-          <input type="text" required></input>
+          <input type="text" required name='username'></input>
           <label>Enter your username</label>
         </div>
         <div class="input-field">
-          <input type="password" required></input>
+          <input type="password" required name='password'></input>
           <label>Enter your password</label>
         </div>
         <button type="submit">Sign up</button>
         <div class="login">
-          <p>Have an account? <a href="#">Log in</a></p>
+          <p>Have an account? <a href="/login">Log in</a></p>
         </div>
       </form>
     </div>
@@ -24,4 +30,5 @@ function Register() {
     )
 }
 
-export default Register
+
+export default Register;
